+++ lib/td_audit_web/views/filters_view.ex
@@ -1,0 +1,7 @@
+defmodule TdAuditWeb.FiltersView do
+  use TdAuditWeb, :view
+
+  def render("filters.json", %{filters: filters}) do
+    Map.take(filters, [:template, :content])
+  end
+end